refactor(recall): pay the ratchet raise instead of taking another
`recall.ts` went 739 -> 744 code lines earlier in this release to ship a 5x
embedding fix an operator was already paying for. That raise named its debt
rather than forgetting it. This is the debt.

Merge, rank and the two text projections moved to `brain/recall-shape.ts`:
124 lines out, and what stays is the part that genuinely needs a database.

## Why these four and not an arbitrary slice

They are the four functions in the file with real unit coverage — 81
assertions across three suites — and that coverage was proven green against
the ORIGINAL code before the new file existed. So the move needed no
characterization pass written for it; the characterization was already there,
which is exactly the property that makes a refactor safe rather than hopeful.

Nothing about their behaviour changed. Same assertions, same count, new
location.

## The freeze is set to what the file IS

689, not left at 744. A freeze above the real number is a budget for 55 lines
nobody argued for.

## Three gates had to follow the code

`hybrid-retrieval` and `rerank` read `recall.ts` as source to assert on it, and
two of those assertions were about functions that had moved. They now read BOTH
halves — a gate that followed only one would go quietly vacuous the next time a
function moves between them, which is worse than the failure I got.

## What is deliberately left

`recall-shape.ts` imports `RecallResult` and `RecallKnowledgeType` back from
`recall.ts`. That is not a runtime cycle: a type-only import is erased by
TypeScript, so the emitted dependency is one-way. It is still the wrong shape
to keep, and the fix is a leaf types module — which is precisely what
`config/rights-shape.ts` already exists to be, for the same reason. That is a
change to every importer of those types and belongs on its own rather than
bolted onto a file move.
